requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.3
Markdown==2.6.9
pytz==2017.3
selenium==3.7.0
